requirements.txt
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
pandas==1.4.3
requests==2.28.1
black==22.6.0